fix: Split container to several components
기존 setting container가 구조가 난잡하다고 판단하여 여러 개의 컴포넌트로 쪼개었습니다.
